Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/parsing/transition-delay-computed-expected.txt


FAIL Property transition-delay value '-500ms, calc(2 * 3s)' assert_equals: expected "-0.5s, 6s" but got "6s"
